Move RoutingPolicies to routing package
